flatten deploy module per @cdmistman
